requirements.txt
websockets
requests
python-socketio
python-socketio[client]